:sparkles: JSX Automatic Runtime に対応した
