--- packages/build-time/package.json
@@ -1,16 +1,0 @@
-{
-  "private": true,
-  "name": "@frostleaf/build-time",
-  "version": "0.1.0",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "check": "tsc --noEmit",
-    "build": "tsc",
-    "format": "prettier --write ."
-  },
-  "devDependencies": {
-    "prettier": "2.6.2",
-    "typescript": "4.7.2"
-  }
-}
