--- packages/build-time/src/fragment.ts
@@ -1,1 +1,0 @@
-export const Fragment = Symbol("Fragment");
