--- packages/build-time/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./fragment";
-export * from "./h";
-export * from "./jsx";
-export * from "./render";
-export * from "./types";
