+++ packages/core/src/global.d.ts
