+++ packages/example/frostleaf-env.d.ts
@@ -1,1 +1,1 @@
-/// <reference types="@frostleaf/build-time/global" />
+/// <reference types="@frostleaf/core/global" />
