+++ packages/example/pages/index.tsx
@@ -1,5 +1,4 @@
-import * as Frostleaf from "@frostleaf/build-time";
-import { Page } from "@frostleaf/build-time";
+import { Page } from "@frostleaf/core";
 
 import Hello from "../components/Hello";
 import World from "../components/World";
