+++ packages/jsx-runtime/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@frostleaf/jsx-runtime",
+  "version": "0.1.0",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "check": "tsc --noEmit",
+    "build": "tsc"
+  },
+  "dependencies": {
+    "@frostleaf/core": "workspace:../core"
+  },
+  "devDependencies": {
+    "typescript": "4.7.2"
+  }
+}
